+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/metrics/AbstractJavaMetric.java
@@ -23,12 +23,25 @@ import net.sourceforge.pmd.lang.metrics.Metric;
  */
 public abstract class AbstractJavaMetric<N extends Node> implements Metric<N> {
 
+
     protected List<JavaQualifiedName> findAllCalls(ASTMethodOrConstructorDeclaration node) {
         List<JavaQualifiedName> result = new ArrayList<>();
         // TODO:cf findAllCalls
         // Needs TypeRes
         // Find the qualified names of all methods called in that method's block
         return result;
+    }
+
+
+    @Override
+    public final boolean equals(Object o) {
+        return o != null && o.getClass() == this.getClass();
+    }
+
+
+    @Override
+    public final int hashCode() {
+        return getClass().hashCode();
     }
 
 
@@ -39,16 +52,6 @@ public abstract class AbstractJavaMetric<N extends Node> implements Metric<N> {
      */
     protected static JavaSignatureMatcher getSignatureMatcher() {
         return JavaMetrics.getFacade().getTopLevelPackageStats();
-    }
-
-    @Override
-    public final boolean equals(Object o) {
-        return o != null && o.getClass() == this.getClass();
-    }
-
-    @Override
-    public final int hashCode() {
-        return getClass().hashCode();
     }
 
 }