+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/rule/metrics/MetricsRulesTest.java
@@ -28,5 +28,6 @@ public class MetricsRulesTest extends SimpleAggregatorTst {
         addRule(RULESET, "CyclomaticComplexity");
         addRule(RULESET, "NcssCount");
         addRule(RULESET, "NPathComplexity");
+        addRule(RULESET, "DataClass");
     }
 }